Move source files to src and convert to ts modules
