+++ src/rules/padding-before-describe-blocks.ts
@@ -1,29 +1,28 @@
 /**
- * @fileoverview Enforces a single like of padding between test blocks with a describe
- * @author Dan
+ * @fileoverview Enforces single line padding around describe blocks
+ * @author Dan Green-Leipciger
  */
 "use strict";
 
-const { padBefore } = require("../utils");
+import { padBefore } from "../utils";
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
 
 const beforeMessage =
-  "You need a newline or comment before an `it` or `test` block when it comes after another expression";
+  "You need a newline or comment before a describe block when it comes after another expression";
 
-const expressionName = node =>
-  node.expression && node.expression.callee && node.expression.callee.name;
-const isTestBlock = node =>
-  expressionName(node) === "test" || expressionName(node) === "it";
+const isDescribe = node =>
+  node.expression &&
+  node.expression.callee &&
+  node.expression.callee.name === "describe";
 
-module.exports = {
+export default {
   meta: {
     docs: {
-      description:
-        "Enforces at least a line of padding before test blocks within a describe",
-      category: "Formatting",
-      recommended: true
+      description: "Enforces at least a line of padding before describe blocks",
+      category: "Fill me in",
+      recommended: false
     },
     fixable: "whitespace", // or "code" or "whitespace"
     schema: [
@@ -33,13 +32,13 @@ module.exports = {
   beforeMessage,
   create(context) {
     const filePath = context && context.getFilename();
-    const isTestFile =
+    const isTest =
       filePath.includes("test") ||
       filePath.includes("spec") ||
       filePath.includes("<input>");
     return {
       ExpressionStatement(node) {
-        if (!isTestFile || !isTestBlock(node)) {
+        if (!isTest || !isDescribe(node)) {
           return;
         }
         padBefore({ context, node, beforeMessage });
